+++ lib/widgets/note_list_del_button.dart
@@ -26,6 +26,7 @@ library;
 
 import 'package:flutter/material.dart';
 
+import 'package:markdown_tooltip/markdown_tooltip.dart';
 import 'package:solidui/solidui.dart';
 
 import 'package:communitypod/common/rest_api/file_helper.dart';
@@ -42,7 +43,9 @@ import 'package:communitypod/widgets/loading_animation.dart' as loading;
 /// - [selectedNotes] - list of selected notes.
 /// - [childPage] - child widget to return to.
 /// - [scaffoldController] - Controller for the Solid scaffold.
-/// - [isSelectionMode] - flag denoting whether notes were selected
+/// - [isSelectionMode] - flag denoting whether notes were selected.
+/// - [isExtFileSelected] - flag denoting whether an external note
+///  in selection.
 /// - [isExternal] - flag denoting whether note is an external
 /// note shared to the user.
 
@@ -51,6 +54,7 @@ class NoteListDelButton extends StatelessWidget {
   final Widget childPage;
   final SolidScaffoldController scaffoldController;
   final bool isSelectionMode;
+  final bool isExtFileSelected;
   final bool isExternal;
 
   const NoteListDelButton({
@@ -59,6 +63,7 @@ class NoteListDelButton extends StatelessWidget {
     required this.childPage,
     required this.scaffoldController,
     this.isSelectionMode = false,
+    this.isExtFileSelected = false,
     this.isExternal = false,
   });
 
@@ -126,14 +131,22 @@ class NoteListDelButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return (!isExternal)
         ? (isSelectionMode)
-            ? TextButton.icon(
-                icon: const Icon(
-                  Icons.delete,
+            ? MarkdownTooltip(
+                message: isExtFileSelected
+                    ? 'You cannot delete notes owned by someone else. Please remove it from the selection'
+                    : 'Delete selected notes',
+                child: TextButton.icon(
+                  icon: const Icon(
+                    Icons.delete,
+                  ),
+                  label: const Text('Delete'),
+                  onPressed: isExtFileSelected
+                      ? null
+                      : () {
+                          // Show inactive button if external file in the selection
+                          deleteListDialog(context);
+                        },
                 ),
-                label: const Text('Delete'),
-                onPressed: () {
-                  deleteListDialog(context);
-                },
               )
             : ElevatedButton.icon(
                 // Uses Theme elevatedButtonTheme for all properties